unsigned/signed comparison fix in Kernel.cpp, a couple cleanups in DisplayHW.cpp, minor casting fix in DebugMPI.cpp
(SVN Revision 333 from svn://localhost/TRUNK/)

+++ System/Debug/PublicMPI/DebugMPI.cpp
@@ -205,7 +205,7 @@ void CDebugMPI::DebugOutErr( tDebugLevel lvl, tErrType err,
 		errstr = pModule_->ErrorToString(err);
 	else
 	{
-		sprintf(buf, "0x%x", err);
+		sprintf(buf, "0x%x", (unsigned int)err);
 		errstr = buf;
 	}
